2024, more like, 2020 Sonic 4 😎

+++ Sonic4Episode1/AppMain/Types/GMS_BOSS5_PART_ACT_INFO.cs
@@ -1,30 +1,3 @@
-using System;
-using System.Collections;
-using System.Collections.Generic;
-using System.Globalization;
-using System.IO;
-using System.IO.IsolatedStorage;
-using System.Linq;
-using System.Reflection;
-using System.Runtime.InteropServices;
-using System.Text;
-using System.Threading;
-using accel;
-using dbg;
-using er;
-using er.web;
-using gs;
-using gs.backup;
-using Microsoft.Xna.Framework;
-using Microsoft.Xna.Framework.Audio;
-using Microsoft.Xna.Framework.Content;
-using Microsoft.Xna.Framework.Graphics;
-using Microsoft.Xna.Framework.Graphics.PackedVector;
-using Microsoft.Xna.Framework.Input;
-using Microsoft.Xna.Framework.Media;
-using mpp;
-using setting;
-
 public partial class AppMain
 {
     public class GMS_BOSS5_PART_ACT_INFO
@@ -39,8 +12,8 @@ public partial class AppMain
 
         public GMS_BOSS5_PART_ACT_INFO()
         {
-            this.act_id = (ushort)0;
-            this.is_maintain = this.is_repeat = (byte)0;
+            this.act_id = 0;
+            this.is_maintain = this.is_repeat = 0;
             this.mtn_spd = this.blend_spd = 0.0f;
             this.is_blend = this.is_merge_manual = 0;
         }
