+++ Sonic4Episode1/AppMain/Types/GSS_MAIN_SYS_INFO.cs
@@ -1,29 +1,4 @@
-using System;
-using System.Collections;
-using System.Collections.Generic;
-using System.Globalization;
-using System.IO;
-using System.IO.IsolatedStorage;
-using System.Linq;
-using System.Reflection;
-using System.Runtime.InteropServices;
-using System.Text;
-using System.Threading;
-using accel;
-using dbg;
-using er;
-using er.web;
-using gs;
 using gs.backup;
-using Microsoft.Xna.Framework;
-using Microsoft.Xna.Framework.Audio;
-using Microsoft.Xna.Framework.Content;
-using Microsoft.Xna.Framework.Graphics;
-using Microsoft.Xna.Framework.Graphics.PackedVector;
-using Microsoft.Xna.Framework.Input;
-using Microsoft.Xna.Framework.Media;
-using mpp;
-using setting;
 
 public partial class AppMain
 {
@@ -70,7 +45,7 @@ public partial class AppMain
             this.final_clear_count = 0U;
             this.is_save_run = 0U;
             this.Save.Init();
-            this.prev_stage_id = (ushort)0;
+            this.prev_stage_id = 0;
             this.is_spe_clear = false;
             this.is_first_play = false;
         }
