+++ Sonic4Episode1/AppMain/Types/GmkPistonData.cs
@@ -1,30 +1,3 @@
-using System;
-using System.Collections;
-using System.Collections.Generic;
-using System.Globalization;
-using System.IO;
-using System.IO.IsolatedStorage;
-using System.Linq;
-using System.Reflection;
-using System.Runtime.InteropServices;
-using System.Text;
-using System.Threading;
-using accel;
-using dbg;
-using er;
-using er.web;
-using gs;
-using gs.backup;
-using Microsoft.Xna.Framework;
-using Microsoft.Xna.Framework.Audio;
-using Microsoft.Xna.Framework.Content;
-using Microsoft.Xna.Framework.Graphics;
-using Microsoft.Xna.Framework.Graphics.PackedVector;
-using Microsoft.Xna.Framework.Input;
-using Microsoft.Xna.Framework.Media;
-using mpp;
-using setting;
-
 public partial class AppMain
 {
     public class GmkPistonData
@@ -33,25 +6,25 @@ public partial class AppMain
         {
       new short[8]
       {
-        (short) 56,
-        (short) 32,
-        (short) -28,
-        (short) 0,
-        (short) -28,
-        (short) 32,
-        (short) 28,
-        (short) 0
+         56,
+         32,
+         -28,
+         0,
+         -28,
+         32,
+         28,
+         0
       },
       new short[8]
       {
-        (short) 56,
-        (short) 32,
-        (short) -28,
-        (short) -32,
-        (short) -28,
-        (short) 0,
-        (short) 28,
-        (short) 32
+         56,
+         32,
+         -28,
+         -32,
+         -28,
+         0,
+         28,
+         32
       }
         };
         public const int GME_GMK_TYPE_PISTON_UP = 0;
