+++ Sonic4Episode1/AppMain/Types/GmkUpBumperData.cs
@@ -1,38 +1,11 @@
-using System;
-using System.Collections;
-using System.Collections.Generic;
-using System.Globalization;
-using System.IO;
-using System.IO.IsolatedStorage;
-using System.Linq;
-using System.Reflection;
-using System.Runtime.InteropServices;
-using System.Text;
-using System.Threading;
-using accel;
-using dbg;
-using er;
-using er.web;
-using gs;
-using gs.backup;
-using Microsoft.Xna.Framework;
-using Microsoft.Xna.Framework.Audio;
-using Microsoft.Xna.Framework.Content;
-using Microsoft.Xna.Framework.Graphics;
-using Microsoft.Xna.Framework.Graphics.PackedVector;
-using Microsoft.Xna.Framework.Input;
-using Microsoft.Xna.Framework.Media;
-using mpp;
-using setting;
-
 public partial class AppMain
 {
     public class GmkUpBumperData
     {
         public static readonly short[][] tbl_gm_gmk_upbumper_rect = new short[2][]
         {
-      new short[4]{ (short) 0, (short) -8, (short) 8, (short) 0 },
-      new short[4]{ (short) -8, (short) -8, (short) 0, (short) 0 }
+      new short[4]{  0,  -8,  8,  0 },
+      new short[4]{ -8, -8, 0, 0 }
         };
         public const int GME_GMK_TYPE_UPBUMPER_L = 0;
         public const int GME_GMK_TYPE_UPBUMPER_R = 1;
