+++ Sonic4Episode1/gs/SSpecial.cs
@@ -4,9 +4,6 @@
 // MVID: 093CE2FC-33E2-4332-B0FE-1EA1E44D3AE7
 // Assembly location: C:\Users\wamwo\Documents\GitHub\Sonic4Ep1-WP7-Decompilation\XAP\Sonic4 ep I.dll
 
-using System;
-using System.IO;
-
 namespace gs.backup
 {
     public class SSpecial : SBase
@@ -37,15 +34,15 @@ namespace gs.backup
 
         private SSpecialSolo getSpecialSolo(uint index)
         {
-            return stages[(int) index];
+            return stages[(int)index];
         }
 
         private SSpecialSolo getSpecialSolo(SStage.EStage.Type index)
         {
-            return stages[(int) index];
+            return stages[(int)index];
         }
 
-        public SSpecialSolo this[int index] => stages[(int) index];
+        public SSpecialSolo this[int index] => stages[index];
 
         public static uint GetSize()
         {
